+++ source/rok4_tasks/rok4_tasks/manager_based/locomotion/__init__.py
@@ -1,0 +1,1 @@
+"""Locomotion task configurations for RoK4."""